move sortablewidget and widgetwrapper into their own files
